fix(gates): restore raw-handle debt and file-size gates broken by the #7467 batch
Eleven canonical across_* conversions (total 1010 -> 999), a documented
loop-reread exception for enumeration.rs, and an indexing.rs split
(fill_extend.rs) back under the 2000-line cap.
