+++ crates/perry-runtime/src/array/indexing.rs
@@ -1226,233 +1226,6 @@ pub extern "C" fn js_array_set_f64_extend(
     }
 }
 
-/// Bulk numeric dense fill for compiler-proven trivial loops:
-/// `for (let i = 0; i < end; i++) arr[i] = constant`.
-///
-/// This keeps JavaScript semantics for frozen/sealed/descriptor arrays by
-/// falling back to the ordinary extending setter. The fast path is restricted
-/// to dense writes from index 0 through `end - 1`, so the resulting live
-/// payload is entirely numeric and can be marked raw-f64 / pointer-free once.
-#[no_mangle]
-pub extern "C" fn js_array_fill_f64_const_extend(
-    arr: *mut ArrayHeader,
-    end: u32,
-    value: f64,
-) -> *mut ArrayHeader {
-    if end == 0 {
-        return clean_arr_ptr_mut(arr);
-    }
-    let arr = clean_arr_ptr_mut(arr);
-    if arr.is_null() {
-        let new_arr = js_array_alloc(end);
-        return js_array_fill_f64_const_extend(new_arr, end, value);
-    }
-    note_array_index_write(arr as usize);
-    let flags = array_object_flags(arr);
-    if flags
-        & (crate::gc::OBJ_FLAG_FROZEN
-            | crate::gc::OBJ_FLAG_SEALED
-            | crate::gc::OBJ_FLAG_NO_EXTEND
-            | crate::gc::OBJ_FLAG_ARRAY_DESCRIPTORS)
-        != 0
-        || crate::buffer::is_registered_buffer(arr as usize)
-        || crate::typedarray::lookup_typed_array_kind(arr as usize).is_some()
-    {
-        let mut out = arr;
-        for i in 0..end {
-            out = js_array_set_f64_extend(out, i, value);
-        }
-        return out;
-    }
-
-    let Some(number) = value_bits_to_number(value.to_bits()) else {
-        let mut out = arr;
-        for i in 0..end {
-            out = js_array_set_f64_extend(out, i, value);
-        }
-        return out;
-    };
-
-    unsafe {
-        let old_length = (*arr).length;
-        let raw_before = array_numeric_layout(arr) == Some(NumericArrayLayout::RawF64);
-        if old_length > end && !raw_before {
-            let mut fallback = arr;
-            for i in 0..end {
-                fallback = js_array_set_f64_extend(fallback, i, value);
-            }
-            return fallback;
-        }
-
-        let mut out = if end > (*arr).capacity {
-            js_array_grow(arr, end)
-        } else {
-            arr
-        };
-        out = clean_arr_ptr_mut(out);
-        if out.is_null() || end > (*out).capacity {
-            let mut fallback = arr;
-            for i in 0..end {
-                fallback = js_array_set_f64_extend(fallback, i, value);
-            }
-            return fallback;
-        }
-        let elements_ptr = (out as *mut u8).add(std::mem::size_of::<ArrayHeader>()) as *mut f64;
-        for i in 0..end as usize {
-            // GC_STORE_AUDIT(POINTER_FREE): bulk numeric fill writes raw f64s only.
-            ptr::write(elements_ptr.add(i), number);
-        }
-        if (*out).length < end {
-            (*out).length = end;
-        }
-        set_array_numeric_layout(out, NumericArrayLayout::RawF64);
-        out
-    }
-}
-
-/// Bulk numeric dense fill for loops bounded by current array length:
-/// `for (let i = 0; i < arr.length; i++) arr[i] = constant`.
-///
-/// If the receiver is exotic (frozen/sealed/descriptors/etc.), the fallback
-/// re-reads `.length` on every iteration so it preserves the source loop's
-/// observable behavior when setters mutate array length.
-#[no_mangle]
-pub extern "C" fn js_array_fill_f64_const_len_extend(
-    arr: *mut ArrayHeader,
-    value: f64,
-) -> *mut ArrayHeader {
-    let arr = clean_arr_ptr_mut(arr);
-    let end = js_array_length(arr);
-    if end == 0 || arr.is_null() {
-        return arr;
-    }
-    note_array_index_write(arr as usize);
-    let flags = array_object_flags(arr);
-    if flags
-        & (crate::gc::OBJ_FLAG_FROZEN
-            | crate::gc::OBJ_FLAG_SEALED
-            | crate::gc::OBJ_FLAG_NO_EXTEND
-            | crate::gc::OBJ_FLAG_ARRAY_DESCRIPTORS)
-        != 0
-        || crate::buffer::is_registered_buffer(arr as usize)
-        || crate::typedarray::lookup_typed_array_kind(arr as usize).is_some()
-    {
-        let mut out = arr;
-        let mut i = 0;
-        while i < js_array_length(out) {
-            out = js_array_set_f64_extend(out, i, value);
-            i += 1;
-        }
-        return out;
-    }
-    js_array_fill_f64_const_extend(arr, end, value)
-}
-
-/// Bulk numeric dense fill for compiler-proven trivial loops:
-/// `for (let i = 0; i < end; i++) arr[i] = i`.
-#[no_mangle]
-pub extern "C" fn js_array_fill_f64_iota_extend(
-    arr: *mut ArrayHeader,
-    end: u32,
-) -> *mut ArrayHeader {
-    if end == 0 {
-        return clean_arr_ptr_mut(arr);
-    }
-    let arr = clean_arr_ptr_mut(arr);
-    if arr.is_null() {
-        let new_arr = js_array_alloc(end);
-        return js_array_fill_f64_iota_extend(new_arr, end);
-    }
-    note_array_index_write(arr as usize);
-    let flags = array_object_flags(arr);
-    if flags
-        & (crate::gc::OBJ_FLAG_FROZEN
-            | crate::gc::OBJ_FLAG_SEALED
-            | crate::gc::OBJ_FLAG_NO_EXTEND
-            | crate::gc::OBJ_FLAG_ARRAY_DESCRIPTORS)
-        != 0
-        || crate::buffer::is_registered_buffer(arr as usize)
-        || crate::typedarray::lookup_typed_array_kind(arr as usize).is_some()
-    {
-        let mut out = arr;
-        for i in 0..end {
-            out = js_array_set_f64_extend(out, i, i as f64);
-        }
-        return out;
-    }
-
-    unsafe {
-        let old_length = (*arr).length;
-        let raw_before = array_numeric_layout(arr) == Some(NumericArrayLayout::RawF64);
-        if old_length > end && !raw_before {
-            let mut fallback = arr;
-            for i in 0..end {
-                fallback = js_array_set_f64_extend(fallback, i, i as f64);
-            }
-            return fallback;
-        }
-
-        let mut out = if end > (*arr).capacity {
-            js_array_grow(arr, end)
-        } else {
-            arr
-        };
-        out = clean_arr_ptr_mut(out);
-        if out.is_null() || end > (*out).capacity {
-            let mut fallback = arr;
-            for i in 0..end {
-                fallback = js_array_set_f64_extend(fallback, i, i as f64);
-            }
-            return fallback;
-        }
-        let elements_ptr = (out as *mut u8).add(std::mem::size_of::<ArrayHeader>()) as *mut f64;
-        for i in 0..end as usize {
-            // GC_STORE_AUDIT(POINTER_FREE): bulk iota fill writes raw f64s only.
-            ptr::write(elements_ptr.add(i), i as f64);
-        }
-        if (*out).length < end {
-            (*out).length = end;
-        }
-        set_array_numeric_layout(out, NumericArrayLayout::RawF64);
-        out
-    }
-}
-
-/// Bulk numeric dense fill for loops bounded by current array length:
-/// `for (let i = 0; i < arr.length; i++) arr[i] = i`.
-///
-/// If the receiver is exotic (frozen/sealed/descriptors/etc.), the fallback
-/// re-reads `.length` on every iteration so it preserves the source loop's
-/// observable behavior when setters mutate array length.
-#[no_mangle]
-pub extern "C" fn js_array_fill_f64_iota_len_extend(arr: *mut ArrayHeader) -> *mut ArrayHeader {
-    let arr = clean_arr_ptr_mut(arr);
-    let end = js_array_length(arr);
-    if end == 0 || arr.is_null() {
-        return arr;
-    }
-    note_array_index_write(arr as usize);
-    let flags = array_object_flags(arr);
-    if flags
-        & (crate::gc::OBJ_FLAG_FROZEN
-            | crate::gc::OBJ_FLAG_SEALED
-            | crate::gc::OBJ_FLAG_NO_EXTEND
-            | crate::gc::OBJ_FLAG_ARRAY_DESCRIPTORS)
-        != 0
-        || crate::buffer::is_registered_buffer(arr as usize)
-        || crate::typedarray::lookup_typed_array_kind(arr as usize).is_some()
-    {
-        let mut out = arr;
-        let mut i = 0;
-        while i < js_array_length(out) {
-            out = js_array_set_f64_extend(out, i, i as f64);
-            i += 1;
-        }
-        return out;
-    }
-    js_array_fill_f64_iota_extend(arr, end)
-}
-
 /// Try to perform `arr[i] = arr[i] + delta` over a dense numeric window.
 ///
 /// This is intentionally transactional: the first pass validates the actual
@@ -1548,27 +1321,6 @@ pub extern "C" fn js_array_numeric_range_add_len(receiver: f64, start: f64, delt
     array_numeric_range_add_impl(receiver, start, None, delta)
 }
 
-#[cfg(feature = "keepalive-anchors")]
-#[used]
-static KEEP_ARRAY_FILL_F64_CONST_EXTEND: extern "C" fn(
-    *mut ArrayHeader,
-    u32,
-    f64,
-) -> *mut ArrayHeader = js_array_fill_f64_const_extend;
-#[cfg(feature = "keepalive-anchors")]
-#[used]
-static KEEP_ARRAY_FILL_F64_IOTA_EXTEND: extern "C" fn(*mut ArrayHeader, u32) -> *mut ArrayHeader =
-    js_array_fill_f64_iota_extend;
-#[cfg(feature = "keepalive-anchors")]
-#[used]
-static KEEP_ARRAY_FILL_F64_CONST_LEN_EXTEND: extern "C" fn(
-    *mut ArrayHeader,
-    f64,
-) -> *mut ArrayHeader = js_array_fill_f64_const_len_extend;
-#[cfg(feature = "keepalive-anchors")]
-#[used]
-static KEEP_ARRAY_FILL_F64_IOTA_LEN_EXTEND: extern "C" fn(*mut ArrayHeader) -> *mut ArrayHeader =
-    js_array_fill_f64_iota_len_extend;
 #[cfg(feature = "keepalive-anchors")]
 #[used]
 static KEEP_ARRAY_NUMERIC_RANGE_ADD: extern "C" fn(f64, f64, f64, f64) -> i64 =
@@ -1750,11 +1502,11 @@ pub extern "C" fn js_array_get_index_or_string(arr: *const ArrayHeader, idx: f64
             // that evacuates the receiver; `arr` is a bare Rust local.
             let scope = crate::gc::RuntimeHandleScope::new();
             let arr_handle = scope.root_raw_const_ptr(arr);
-            let key_ptr = crate::string::js_string_from_bytes(key.as_ptr(), key.len() as u32);
-            return array_get_property_by_key(
-                arr_handle.get_raw_const_ptr::<ArrayHeader>(),
-                key_ptr,
-            );
+            // Allocating key build + receiver re-read as one combinator (#7341).
+            let (key_ptr, arr_now) = arr_handle.across_const::<ArrayHeader, _>(|| {
+                crate::string::js_string_from_bytes(key.as_ptr(), key.len() as u32)
+            });
+            return array_get_property_by_key(arr_now, key_ptr);
         }
     }
